+++ __tests__/categories/js/eslint/eslint-test.utils.ts
@@ -3,8 +3,9 @@ import { Config } from 'src/categories/js/eslint/config/config.interface';
 export const createConfig = ({
   dependencies = [],
   eslintConfig = {},
+  ignore = [],
   scripts = [],
   mutations = [],
 }: Partial<Config> = {}) => {
-  return { dependencies, eslintConfig, scripts, mutations };
+  return { dependencies, eslintConfig, ignore, scripts, mutations };
 };