+++ src/categories/js/eslint/eslint.const.ts
@@ -2,3 +2,4 @@ export const SCRIPT_NAME = 'eslint';
 export const PACKAGE_NAME = SCRIPT_NAME;
 export const CLI_NAME = PACKAGE_NAME;
 export const CONFIG_FILE_NAME = '.eslintrc.json';
+export const CONFIG_IGNORE_FILE_NAME = '.eslintignore';