+++ src/categories/js/jest/config/node:ts.config.ts
@@ -3,13 +3,14 @@ import { scripts } from './default.config';
 import { Config } from './config.interface';
 
 const configFileContent = readableMultilineString`
-  /** @type {import('ts-jest/dist/types').InitialOptionsTsJest} */
+  /** @type {import('ts-jest').JestConfigWithTsJest} */
   module.exports = {
     preset: 'ts-jest',
     testEnvironment: 'node',
     moduleDirectories: ['<rootDir>', 'node_modules'],
     testRegex: '.*\\.(spec|test)\\.ts$',
     collectCoverageFrom: ['src/**/*.ts'],
+    passWithNoTests: true,
   };
 `;